pyproject.toml
[project]
name = "testingpython"
version = "0.1.0"
description = "CRM Django Application with React Frontend"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.3",
    "django-environ>=0.12.0",
    "django-filter>=24.3",
    "djangorestframework>=3.16.0",
    "django-cors-headers>=4.5.0",
    "faker>=37.4.0",
    "gunicorn>=23.0.0",
    "psycopg[binary]>=3.2.9",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["testingpython", "api"]
